Refactor: rename http package to httpclient to avoid naming conflicts
