+++ internal/infrastructure/httpclient/client_factory_test.go
@@ -1,4 +1,4 @@
-package http
+package httpclient
 
 import (
 	"io"
