+++ internal/infrastructure/httpclient/errors_test.go
@@ -1,4 +1,4 @@
-package http
+package httpclient
 
 import (
 	"errors"
