+++ internal/infrastructure/httpclient/send_request.go
@@ -1,4 +1,4 @@
-package http
+package httpclient
 
 import (
 	"bytes"
