+++ internal/infrastructure/httpclient/send_request_test.go
@@ -1,4 +1,4 @@
-package http
+package httpclient
 
 import (
 	"fmt"
